+++ packages/select/index.ts
@@ -1,0 +1,8 @@
+import Select from "./src/index.vue";
+export const SelectName = "VfSelect";
+
+(Select as any).install = function (Vue: any) {
+  Vue.component(SelectName, Select);
+};
+
+export default Select;