+++ types/select.d.ts
@@ -1,0 +1,3 @@
+import { VFUIComponent } from "./component";
+
+export declare class VfSelect extends VFUIComponent {}